Lagt till metoder och filter för bokning, ändrat HCP i person från int till double, lagt till metod för bokning i postgress (inte helt färdigställd)

+++ Golf2/Golf2/person.cs
@@ -10,7 +10,7 @@ public class Person
     public string Email { get; set; }
     public string Postalcode { get; set; }
     public string City { get; set; }
-    public int Hcp { get; set; }
+    public double Hcp { get; set; }
     public string Adress { get; set; }
 
     public int TypId { get; set; }
